remove redundant casts due to Configuration#getInt returning int now

+++ graylog2-inputs/src/main/java/org/graylog2/inputs/transports/AmqpTransport.java
@@ -137,11 +137,11 @@ public class AmqpTransport extends ThrottleableTransport {
     public void launch(MessageInput input) throws MisfireException {
         consumer = new AmqpConsumer(
                 configuration.getString(CK_HOSTNAME),
-                (int) configuration.getInt(CK_PORT),
+                configuration.getInt(CK_PORT),
                 configuration.getString(CK_VHOST),
                 configuration.getString(CK_USERNAME),
                 configuration.getString(CK_PASSWORD),
-                (int) configuration.getInt(CK_PREFETCH),
+                configuration.getInt(CK_PREFETCH),
                 configuration.getString(CK_QUEUE),
                 configuration.getString(CK_EXCHANGE),
                 configuration.getString(CK_ROUTING_KEY),
